deployment tested and some bugs fixed

+++ api/deploy.sh
@@ -22,19 +22,19 @@ if [ -z "$USE_LLM" ]; then
   exit 1
 fi
 
-# Build docker-compose first (stored as docker-compose.final.yml)
-docker compose -f docker-compose.build-compose.yml up
-EXIT_CODE=$?
-docker compose -f docker-compose.build-compose.yml down
-
-if [ $EXIT_CODE -ne 0 ]; then
-    echo "Failed to build a docker compose"
-    exit 1
-fi
-
 # What operation we wish to perform
 COMPOSE_COMMAND="up -d --build"
-if [ "$1" == "--stop" ]; then  
+if [ "$#" -eq 0 ]; then
+  # Build docker-compose first (stored as docker-compose.final.yml)
+  docker compose -f docker-compose.build-compose.yml up
+  EXIT_CODE=$?
+  docker compose -f docker-compose.build-compose.yml down
+
+  if [ $EXIT_CODE -ne 0 ]; then
+    echo "Failed to build a docker compose"
+    exit 1
+  fi
+elif [ "$1" == "--stop" ]; then  
   COMPOSE_COMMAND="stop"
 elif [ "$1" == "--remove" ]; then  
   COMPOSE_COMMAND="down"
